+++ keyhint/__init__.py
@@ -1,4 +1,1 @@
-from keyhint import config, context, sheets, utils
-
 __version__ = "0.4.4"
-__all__ = ["config", "context", "sheets", "utils", "__version__"]