+++ keyhint/window.py
@@ -15,20 +15,20 @@ Naming Hierarchy:
 """
 
 import logging
-import re
-from collections.abc import Callable
+import subprocess
+import textwrap
 from pathlib import Path
-from typing import Any
 
-from gi.repository import Adw, Gdk, Gio, GObject, Gtk
+from gi.repository import Adw, Gdk, Gio, GLib, GObject, Gtk, Pango
 
-import keyhint
+from keyhint import __version__, config, context, sheets, utils
 
-logger = logging.getLogger(__name__)
+logger = logging.getLogger("keyhint")
 
 RESOURCE_PATH = Path(__file__).parent / "resources"
 
 
+# TODO: Flatpak
 class BindingsRow(GObject.Object):
     shortcut: str
     label: str
@@ -41,26 +41,27 @@ class BindingsRow(GObject.Object):
         self.filter_text = f"{shortcut} {label} {section}"
 
 
-@Gtk.Template(filename=f"{RESOURCE_PATH}/header_bar.ui")
-class HeaderBar(Gtk.Box):
-    __gtype_name__ = "header_bar"
+@Gtk.Template(filename=f"{RESOURCE_PATH}/headerbar.ui")
+class HeaderBarBox(Gtk.Box):
+    __gtype_name__ = "headerbar_box"
 
     sheet_drop_down: Gtk.DropDown = Gtk.Template.Child()
     search_entry: Gtk.SearchEntry = Gtk.Template.Child()
     fullscreen_button = Gtk.Template.Child()
-    adwheaderbar: Adw.HeaderBar = Gtk.Template.Child()
-    sort_by_size_button = Gtk.Template.Child()
-    sort_by_title_button = Gtk.Template.Child()
-    scroll_vertical_button = Gtk.Template.Child()
-    scroll_horizontal_button = Gtk.Template.Child()
+    headerbar: Adw.HeaderBar = Gtk.Template.Child()
     zoom_spin_button = Gtk.Template.Child()
     fallback_sheet_entry: Gtk.Entry = Gtk.Template.Child()
 
-    def __init__(self, visible: bool = True, decoration: str | None = None) -> None:
+    def __init__(
+        self, for_fullscreen: bool = False, decoration: str | None = None
+    ) -> None:
         super().__init__()
-        self.set_visible(visible)
-        if decoration:
-            self.adwheaderbar.set_decoration_layout(decoration)
+        if for_fullscreen:
+            self.headerbar.set_decoration_layout(":minimize,close")
+            self.fullscreen_button.set_icon_name("view-restore-symbolic")
+            self.set_visible(False)
+        else:
+            self.set_visible(True)
 
 
 @Gtk.Template(filename=f"{RESOURCE_PATH}/window.ui")
@@ -69,14 +70,7 @@ class KeyhintWindow(Gtk.ApplicationWindow):
 
     __gtype_name__ = "main_window"
 
-    sheets: list[dict] = keyhint.sheets.load_sheets()
-    dialog_is_open: bool = False
-    wm_class: str = ""
-    window_title: str = ""
-    sheet_id: str = ""
-    ignore_search_changed: bool = False
-    search_text: str = ""
-
+    overlay: Adw.ToastOverlay = Gtk.Template.Child()
     scrolled_window: Gtk.ScrolledWindow = Gtk.Template.Child()
     container: Gtk.Box = Gtk.Template.Child()
     sheet_container_box: Gtk.FlowBox = Gtk.Template.Child()
@@ -86,21 +80,39 @@ class KeyhintWindow(Gtk.ApplicationWindow):
     def __init__(self, cli_args: dict) -> None:
         """Initialize during window creation."""
         super().__init__()
-        logger.debug("Loaded %s cheatsheets.", len(self.sheets))
 
         self.cli_args = cli_args
-        self.config = keyhint.config.load()
+        self.config = config.load()
+        self.sheets = sheets.load_sheets()
+        self.wm_class, self.window_title = context.detect_active_window()
 
-        self.load_css()
+        self.sheet_id: str = ""
+        self.skip_search_changed: bool = False
+        self.search_text: str = ""
+
+        self.css_provider = utils.create_css_provider(
+            display=self.get_display(), css=f"{RESOURCE_PATH}/style.css"
+        )
+        self.zoom_css_provider = utils.create_css_provider(display=self.get_display())
         self.set_icon_name("keyhint")
 
         # Add normal header bar
-        self.headerbar = HeaderBar()
+        self.headerbar = HeaderBarBox()
         self.set_titlebar(self.headerbar)
 
         # Add second header bar which is only visible in fullscreen mode
-        self.headerbar_fs = HeaderBar(decoration=":minimize,close", visible=False)
+        self.headerbar_fs = HeaderBarBox(for_fullscreen=True)
         self.container.prepend(self.headerbar_fs)
+
+        self.shortcut_column_factory = Gtk.SignalListItemFactory()
+        self.shortcut_column_factory.connect("bind", self.bind_shortcuts_callback)
+
+        self.label_column_factory = Gtk.SignalListItemFactory()
+        self.label_column_factory.connect("bind", self.bind_labels_callback)
+
+        self.list_view_filter = Gtk.CustomFilter.new(
+            match_func=self.list_view_match_func
+        )
 
         # Set up the controls
         self.init_signals()
@@ -111,8 +123,10 @@ class KeyhintWindow(Gtk.ApplicationWindow):
         self.init_event_controller_key()
 
         # Apply loaded configuration
-        self.toggle_orientation(self.config["main"].get("orientation", "horizontal"))
-        self.toggle_fullscreen(self.config["main"].getboolean("fullscreen", False))
+        self.change_zoom(self.config["main"].getint("zoom", 100))
+        self.headerbar.fallback_sheet_entry.set_text(
+            self.config["main"].get("fallback_cheatsheet", "keyhint")
+        )
 
         # Make sure the window is focused
         self.present()
@@ -120,14 +134,6 @@ class KeyhintWindow(Gtk.ApplicationWindow):
 
     def init_signals(self) -> None:
         """Connect signals to methods."""
-        # Fullscreen button
-        self.headerbar.fullscreen_button.connect(
-            "clicked", lambda _: self.toggle_fullscreen()
-        )
-        self.headerbar_fs.fullscreen_button.connect(
-            "clicked", lambda _: self.toggle_fullscreen()
-        )
-
         # Search entry
         self.search_changed_handler = self.headerbar.search_entry.connect(
             "search-changed", self.on_search_entry_changed
@@ -141,8 +147,23 @@ class KeyhintWindow(Gtk.ApplicationWindow):
             "notify::selected-item", self.on_sheet_drop_down_changed
         )
 
+        # Zoom
+        self.headerbar.zoom_spin_button.connect(
+            "value-changed", lambda btn: self.change_zoom(btn.get_value())
+        )
+        self.headerbar_fs.zoom_spin_button.connect(
+            "value-changed", lambda btn: self.change_zoom(btn.get_value())
+        )
+
+        # Default sheet
+        self.headerbar.fallback_sheet_entry.connect(
+            "changed",
+            lambda entry: self.change_default_sheet(entry.get_text()),
+        )
+
         # Sheets filter
-        self.sheet_container_box.set_filter_func(self.filter_sections)
+        self.sheet_container_box.set_filter_func(self.filter_sections_func)
+        self.sheet_container_box.set_sort_func(self.sort_sections_func)
 
         # Window state
         self.connect("notify::fullscreened", self.on_fullscreen_state_changed)
@@ -151,6 +172,44 @@ class KeyhintWindow(Gtk.ApplicationWindow):
         """Add actions used by main menu."""
         action = Gio.SimpleAction.new("about", None)
         action.connect("activate", self.on_about_action)
+        self.add_action(action)
+
+        action = Gio.SimpleAction.new("debug_info", None)
+        action.connect("activate", self.on_debug_action)
+        self.add_action(action)
+
+        action = Gio.SimpleAction.new("open_folder", None)
+        action.connect("activate", self.on_open_folder_action)
+        self.add_action(action)
+
+        action = Gio.SimpleAction.new("create_new_sheet", None)
+        action.connect("activate", self.on_create_new_sheet)
+        self.add_action(action)
+
+        action = Gio.SimpleAction.new_stateful(
+            name="fullscreen",
+            state=GLib.Variant(
+                "b", self.config["main"].getboolean("fullscreen", False)
+            ),
+            parameter_type=None,
+        )
+        action.connect("activate", self.toggle_fullscreen)
+        self.add_action(action)
+
+        action = Gio.SimpleAction.new_stateful(
+            name="sort_by",
+            state=GLib.Variant("s", self.config["main"].get("sort_by", "size")),
+            parameter_type=GLib.VariantType.new("s"),
+        )
+        action.connect("activate", self.toggle_sort)
+        self.add_action(action)
+
+        action = Gio.SimpleAction.new_stateful(
+            name="orientation",
+            state=GLib.Variant("s", self.config["main"].get("orientation", "vertical")),
+            parameter_type=GLib.VariantType.new("s"),
+        )
+        action.connect("activate", self.toggle_orientation)
         self.add_action(action)
 
     def init_event_controller_key(self) -> None:
@@ -177,7 +236,7 @@ class KeyhintWindow(Gtk.ApplicationWindow):
         self.headerbar.sheet_drop_down.set_selected(select_idx)
 
     def init_search_entry(self) -> None:
-        """Let search entry capture key events."""
+        """Let search entries capture key events."""
         # Bind changes between search entries in header bar and for fullscreen
         evk = Gtk.EventControllerKey()
         evk.connect("key-pressed", self.on_search_entry_key_pressed)
@@ -191,10 +250,6 @@ class KeyhintWindow(Gtk.ApplicationWindow):
     def init_binds(self) -> None:
         """Bind properties of controls between both header bars."""
         widgets = [
-            ("sort_by_size_button", "active"),
-            ("sort_by_title_button", "active"),
-            ("scroll_vertical_button", "active"),
-            ("scroll_horizontal_button", "active"),
             ("zoom_spin_button", "value"),
             ("fallback_sheet_entry", "text"),
             ("sheet_drop_down", "selected"),
@@ -207,43 +262,59 @@ class KeyhintWindow(Gtk.ApplicationWindow):
                 GObject.BindingFlags.BIDIRECTIONAL,
             )
 
-    def on_sheet_drop_down_changed(
-        self, drop_down: Gtk.DropDown, _: GObject.Parameter
-    ) -> None:
-        """Execute on change of the sheet selection dropdown."""
-        if not (selected_item := drop_down.get_selected_item()):
-            return
+    def change_default_sheet(self, sheet_id: str) -> None:
+        """Set the default sheet to use."""
+        if sheet_id not in [s["id"] for s in self.sheets]:
+            sheet_id = "keyhint"
+            css_class = "error"
+        else:
+            css_class = ""
+        self.headerbar.fallback_sheet_entry.set_css_classes([css_class])
+        self.headerbar_fs.fallback_sheet_entry.set_css_classes([css_class])
+        self.config.set_persistent("main", "fallback_cheatsheet", sheet_id)
 
-        self.sheet_id = selected_item.get_string()
-        self.populate_sheet_container()
+    def change_zoom(self, value: float) -> None:
+        """Set the zoom level of the sheet container."""
+        self.headerbar.zoom_spin_button.set_value(value)
+        self.zoom_css_provider.load_from_string(
+            f"""
+            .sheet_container_box,
+            .sheet_container_box .bindings-section header label {{
+                font-size: {value}%;
+            }}
+            """
+        )
+        self.config.set_persistent("main", "zoom", str(int(value)))
 
-    def toggle_orientation(self, orientation: str | None = None) -> None:
+    def toggle_orientation(self, action: Gio.SimpleAction, state: GLib.Variant) -> None:
         """Set the orientation of the sheet container."""
-        if not orientation:
-            orientation = str(self.sheet_container_box.get_orientation().value_nick)
+        action.set_state(state)
 
         # The used GTK orientation is the opposite of the config value, which follows
         # the naming from user perspective!
+        value = str(state).strip("'")
         gtk_orientation = (
             Gtk.Orientation.VERTICAL
-            if orientation == "horizontal"
+            if value == "horizontal"
             else Gtk.Orientation.HORIZONTAL
         )
+
         self.sheet_container_box.set_orientation(gtk_orientation)
-        self.config.set_persistent("main", "orientation", orientation)
+        self.config.set_persistent("main", "orientation", value)
 
-    def toggle_fullscreen(self, to_fullscreen: bool | None = None) -> None:
+    def toggle_sort(self, action: Gio.SimpleAction, state: GLib.Variant) -> None:
+        """Set the order of the sections."""
+        action.set_state(state)
+
+        self.config.set_persistent("main", "sort_by", str(state).strip("'"))
+        self.sheet_container_box.invalidate_sort()
+
+    def toggle_fullscreen(self, action: Gio.SimpleAction, state: GLib.Variant) -> None:
         """Set the fullscreen state."""
-        if to_fullscreen is self.is_fullscreen():
-            return
+        to_fullscreen = not bool(action.get_state())
+        action.set_state(GLib.Variant("b", to_fullscreen))
 
-        if to_fullscreen is None:
-            to_fullscreen = not self.is_fullscreen()
-
-        self.headerbar.fullscreen_button.set_active(to_fullscreen)
-        self.headerbar_fs.fullscreen_button.set_active(to_fullscreen)
-
-        self.ignore_search_changed = True
+        self.skip_search_changed = True
         if to_fullscreen:
             self.headerbar_fs.search_entry.set_text(self.search_text)
             self.fullscreen()
@@ -253,13 +324,18 @@ class KeyhintWindow(Gtk.ApplicationWindow):
 
         self.config.set_persistent("main", "fullscreen", to_fullscreen)
 
-    def on_fullscreen_state_changed(self, _: Gtk.Widget, __: GObject.Parameter) -> None:
-        """Hide header bar in title and show it in window instead, and vice versa."""
-        if self.is_fullscreen():
-            self.headerbar_fs.set_visible(True)
+    def scroll(self, to_start: bool, by_page: bool) -> None:
+        if self.sheet_container_box.get_orientation() == 1:
+            adj = self.scrolled_window.get_hadjustment()
         else:
-            self.headerbar_fs.set_visible(False)
-        self.focus_search_entry()
+            adj = self.scrolled_window.get_vadjustment()
+
+        default_distance = 25
+        distance = adj.get_page_size() if by_page else default_distance
+        if to_start:
+            distance *= -1
+
+        adj.set_value(adj.get_value() + distance)
 
     def focus_search_entry(self) -> None:
         """Focus search entry depending on state."""
@@ -270,10 +346,25 @@ class KeyhintWindow(Gtk.ApplicationWindow):
             self.headerbar.search_entry.grab_focus()
             self.headerbar.search_entry.set_position(-1)
 
+    def show_create_new_sheet_toast(self) -> None:
+        toast = Adw.Toast.new(f"No cheatsheet found for '{self.wm_class}'.")
+        toast.set_button_label("Create new")
+        toast.set_action_name("win.create_new_sheet")
+        toast.set_timeout(5)
+        self.overlay.add_toast(toast)
+
+    def on_fullscreen_state_changed(self, _: Gtk.Widget, __: GObject.Parameter) -> None:
+        """Hide header bar in title and show it in window instead, and vice versa."""
+        if self.is_fullscreen():
+            self.headerbar_fs.set_visible(True)
+        else:
+            self.headerbar_fs.set_visible(False)
+        self.focus_search_entry()
+
     def on_search_entry_changed(self, search_entry: Gtk.SearchEntry) -> None:
         """Execute on change of the sheet selection dropdown."""
-        if self.ignore_search_changed:
-            self.ignore_search_changed = False
+        if self.skip_search_changed:
+            self.skip_search_changed = False
             return
 
         if search_entry.get_text() == self.search_text:
@@ -293,6 +384,16 @@ class KeyhintWindow(Gtk.ApplicationWindow):
         if keycode == Gdk.KEY_Escape:
             self.close()
 
+    def on_sheet_drop_down_changed(
+        self, drop_down: Gtk.DropDown, _: GObject.Parameter
+    ) -> None:
+        """Execute on change of the sheet selection dropdown."""
+        if not (selected_item := drop_down.get_selected_item()):
+            return
+
+        self.sheet_id = selected_item.get_string()
+        self.populate_sheet_container()
+
     def on_key_pressed(
         self,
         evk: Gtk.EventControllerKey,
@@ -300,57 +401,94 @@ class KeyhintWindow(Gtk.ApplicationWindow):
         keyval,  # noqa: ANN001
         modifier,  # noqa: ANN001
     ) -> None:
-        match keycode:
-            case Gdk.KEY_Escape:
+        ctrl_pressed = modifier == Gdk.ModifierType.CONTROL_MASK
+        match keycode, ctrl_pressed:
+            case Gdk.KEY_Escape, _:
                 # If the dialog is open, that one is closed. Otherwise close the window
-                if self.dialog_is_open:
-                    self.dialog_is_open = False
+                self.close()
+            case Gdk.KEY_F11, _:
+                self.activate_action("win.fullscreen")
+            case Gdk.KEY_f, True:
+                # Focus search entry
+                if self.is_fullscreen():
+                    self.headerbar_fs.search_entry.grab_focus()
                 else:
-                    self.close()
-            case Gdk.KEY_F11:
-                self.toggle_fullscreen()
-            case Gdk.KEY_f:
-                if modifier == Gdk.ModifierType.CONTROL_MASK:
-                    # Focus search entry
-                    if self.is_fullscreen():
-                        self.search_entry_fs.grab_focus()
-                    else:
-                        self.search_entry.grab_focus()
-            case Gdk.KEY_s:
-                if modifier == Gdk.ModifierType.CONTROL_MASK:
-                    # Open sheet dropdown
-                    self.sheet_drop_down.grab_focus()
-            case Gdk.KEY_Up | Gdk.KEY_k:
+                    self.headerbar.search_entry.grab_focus()
+            case Gdk.KEY_s, True:
+                # Open sheet dropdown
+                if self.is_fullscreen():
+                    self.headerbar_fs.sheet_drop_down.grab_focus()
+                else:
+                    self.headerbar.sheet_drop_down.grab_focus()
+            case (Gdk.KEY_Up, _) | (Gdk.KEY_k, True):
                 self.scroll(to_start=True, by_page=False)
-            case Gdk.KEY_Down | Gdk.KEY_j:
+            case (Gdk.KEY_Down, _) | (Gdk.KEY_j, True):
                 self.scroll(to_start=False, by_page=False)
-            case Gdk.KEY_Left | Gdk.KEY_h:
-                self.scroll(to_start=True, by_page=False)
-            case Gdk.KEY_Right | Gdk.KEY_l:
-                self.scroll(to_start=False, by_page=False)
-            case Gdk.KEY_Page_Up:
+            case Gdk.KEY_Page_Up, _:
                 self.scroll(to_start=True, by_page=True)
-            case Gdk.KEY_Page_Down:
+            case Gdk.KEY_Page_Down, _:
                 self.scroll(to_start=False, by_page=True)
 
     def on_about_action(self, _: Gio.SimpleAction, __: None) -> None:
         """Execute on click "about" in application menu."""
-        self.dialog_is_open = True
         logo = Gtk.Image.new_from_file(f"{RESOURCE_PATH}/keyhint_icon.svg")
         Gtk.AboutDialog(
             program_name="KeyHint",
-            comments="Cheatsheet for keyboard shortcuts &amp; commands",
-            version=keyhint.__version__,
-            website_label="Website",
+            comments="Cheatsheet for keyboard shortcuts & commands",
+            version=__version__,
+            website_label="Github",
             website="https://github.com/dynobo/keyhint",
             logo=logo.get_paintable(),
-            system_information=self.get_debug_info(),
+            license_type=Gtk.License.MIT_X11,
             modal=True,
             resizable=True,
             transient_for=self,
         ).show()
 
-    def filter_sections(self, child: Gtk.Widget) -> bool:
+    def on_debug_action(self, _: Gio.SimpleAction, __: None) -> None:
+        dialog = Gtk.Dialog(title="Debug Info", transient_for=self, modal=True)
+        label = Gtk.Label()
+        label.set_use_markup(True)
+        label.set_wrap(True)
+        label.set_wrap_mode(Pango.WrapMode.WORD_CHAR)
+        label.set_markup(self.get_debug_info_text())
+        label.set_margin_start(24)
+        label.set_margin_end(24)
+
+        dialog.set_child(label)
+        dialog.show()
+
+    def on_create_new_sheet(self, _: Gio.SimpleAction, __: None) -> None:
+        title = self.wm_class.lower().replace(" ", "")
+        pad = 26 - len(title)
+        template = f"""\
+            id = "{title}"{" " * pad } # Unique ID, used e.g. in cheatsheet dropdown
+            url = ""                       # (Optional) URL to keybinding docs
+
+            [match]
+            regex_wmclass = "{self.wm_class}"
+            regex_title = ".*"                # Narrow down by window title if needed
+
+            [section]
+
+            [section."My Section Title"]      # Add as many sections you like ...
+            "Ctrl + c" = "Copy to clipboard"  # ... with keybinding + description
+            "Ctrl + v" = "Paste from clipboard"
+        """
+        template = textwrap.dedent(template)
+        new_file = config.CONFIG_PATH / f"{title}.toml"
+        for idx in range(1, 100):
+            if new_file.exists():
+                new_file = new_file.with_name(f"{title}_{idx}.toml")
+            else:
+                break
+        new_file.write_text(template)
+        subprocess.Popen(["xdg-open", str(new_file.resolve())])  # noqa: S603, S607
+
+    def on_open_folder_action(self, _: Gio.SimpleAction, __: None) -> None:
+        subprocess.Popen(["xdg-open", str(config.CONFIG_PATH.resolve())])  # noqa: S603, S607
+
+    def filter_sections_func(self, child: Gtk.Widget) -> bool:
         """Filter binding sections based on the search entry text."""
         # If no text, show all sections
         if not self.search_text:
@@ -362,85 +500,52 @@ class KeyhintWindow(Gtk.ApplicationWindow):
         filter_model = selection.get_model()
         return filter_model.get_n_items() > 0
 
-    def scroll(self, to_start: bool, by_page: bool) -> None:
-        if self.sheet_container_box.get_orientation() == 1:
-            adj = self.scrolled_window.get_hadjustment()
-        else:
-            adj = self.scrolled_window.get_vadjustment()
-
-        default_distance = 25
-        distance = adj.get_page_size() if by_page else default_distance
-        if to_start:
-            distance *= -1
-
-        adj.set_value(adj.get_value() + distance)
-
-    def get_sheet_by_id(self, sheet_id: str) -> dict[str, Any]:
-        return next(sheet for sheet in self.sheets if sheet["id"] == sheet_id)
-
-    def get_sheet_id_by_active_window(self) -> str | None:
-        self.wm_class, self.window_title = keyhint.context.detect_active_window()
-
-        matching_sheets = [
-            h
-            for h in self.sheets
-            if re.search(h["match"]["regex_process"], self.wm_class, re.IGNORECASE)
-            and re.search(h["match"]["regex_title"], self.window_title, re.IGNORECASE)
-        ]
-
-        if not matching_sheets:
-            return None
-
-        # First sort by secondary criterion
-        matching_sheets.sort(key=lambda h: len(h["match"]["regex_title"]), reverse=True)
-
-        # Then sort by primary criterion
-        matching_sheets.sort(
-            key=lambda h: len(h["match"]["regex_process"]), reverse=True
-        )
-
-        # First element is (hopefully) the best fitting sheet id
-        return matching_sheets[0]["id"]
+    def sort_sections_func(self, child_a: Gtk.Widget, child_b: Gtk.Widget) -> bool:
+        sort_by = self.config["main"].get("sort_by", "size")
+        if sort_by == "size":
+            return (
+                child_a.get_child().get_model().get_n_items()
+                < child_b.get_child().get_model().get_n_items()
+            )
+        if sort_by == "title":
+            return (
+                child_a.get_child().get_columns().get_item(1).get_title()
+                > child_b.get_child().get_columns().get_item(1).get_title()
+            )
+        return child_a.get_name() > child_b.get_name()
 
     def get_appropriate_sheet_id(self) -> str | None:
         sheet_id = None
 
-        # If sheet-id was provided by option, use that one:
-        if "cheatsheet" in self.cli_args:
-            sheet_id = self.cli_args["cheatsheet"]
-            logger.debug("Using provided sheet-id: %s", sheet_id)
+        # If sheet-id was provided via cli option, use that one
+        if sheet_id := self.cli_args.get("cheatsheet", None):
+            logger.debug("Using provided sheet-id: %s.", sheet_id)
+            return sheet_id
 
         # Else try to find cheatsheet for active window
-        if not sheet_id:
-            sheet_id = self.get_sheet_id_by_active_window()
-            logger.debug("Found matching sheets %s", sheet_id)
+        if sheet_id := sheets.get_sheet_id_by_active_window(
+            sheets=self.sheets, wm_class=self.wm_class, window_title=self.window_title
+        ):
+            logger.debug("Found matching sheet: %s.", sheet_id)
+            return sheet_id
 
-        # First fallback to cli provided default
-        if (not sheet_id) and ("default-cheatsheet" in self.cli_args):
-            sheet_id = self.cli_args["default-cheatsheet"]
-            logger.debug("Using provided default sheet-id: %s", sheet_id)
+        self.show_create_new_sheet_toast()
 
-        # Last fallback to first entry in list
-        if not sheet_id:
-            model = self.sheet_drop_down.get_model()
-            sheet_id = model.get_item(0).get_string()
-            logger.debug("No matching sheet found. Using first sheet in list.")
+        # Use configured fallback
+        if sheet_id := self.config["main"].get("fallback_cheatsheet", ""):
+            logger.debug("Using provided fallback sheet-id: %s.", sheet_id)
+            return sheet_id
 
+        # Fallback to first entry in list
+        model = self.sheet_drop_down.get_model()
+        sheet_id = model.get_item(0).get_string()
+        logger.debug("No matching or fallback sheet found. Using first sheet.")
         return sheet_id
-
-    def load_css(self) -> None:
-        """Load custom global CSS."""
-        css_path = f"{RESOURCE_PATH}/style.css"
-        provider = Gtk.CssProvider()
-        provider.load_from_path(css_path)
-        Gtk.StyleContext().add_provider_for_display(
-            self.get_display(), provider, Gtk.STYLE_PROVIDER_PRIORITY_USER
-        )
 
     def bind_shortcuts_callback(
         self, _: Gtk.SignalListItemFactory, item: Gtk.Widget
     ) -> None:
-        shortcut = keyhint.utils.create_shortcut(item.get_item().shortcut)
+        shortcut = utils.create_shortcut(item.get_item().shortcut)
         self.max_shortcut_width = max(
             self.max_shortcut_width, shortcut.get_preferred_size().natural_size.width
         )
@@ -453,7 +558,9 @@ class KeyhintWindow(Gtk.ApplicationWindow):
         if item.get_item().shortcut:
             child = Gtk.Label(label=label, xalign=0.0)
         else:
-            child = keyhint.utils.create_section_title(label)
+            # section title
+            child = Gtk.Label(xalign=0.0)
+            child.set_markup(f"<b>{label}</b>")
         item.set_child(child)
 
     def list_view_match_func(self, bindings_row: BindingsRow) -> bool:
@@ -461,38 +568,7 @@ class KeyhintWindow(Gtk.ApplicationWindow):
             return self.search_text.lower() in bindings_row.filter_text.lower()
         return True
 
-    def create_list_item_factory(self, callback: Callable) -> Gtk.SignalListItemFactory:
-        factory = Gtk.SignalListItemFactory()
-        factory.connect("bind", callback)
-        return factory
-
-    def create_column_view_column(
-        self,
-        title: str,
-        factory: Gtk.SignalListItemFactory,
-        fixed_width: int | None = None,
-    ) -> Gtk.ColumnViewColumn:
-        column = Gtk.ColumnViewColumn(title=title, factory=factory)
-        if fixed_width:
-            column.set_fixed_width(fixed_width)
-        return column
-
-    def create_column_view(
-        self,
-        selection: Gtk.SelectionModel,
-        shortcut_column: Gtk.ColumnViewColumn,
-        label_column: Gtk.ColumnViewColumn,
-    ) -> Gtk.ColumnView:
-        column_view = Gtk.ColumnView()
-        column_view.get_style_context().add_class("bindings-section")
-        column_view.set_hexpand(True)
-        column_view.set_model(selection)
-        column_view.append_column(shortcut_column)
-        column_view.append_column(label_column)
-        return column_view
-
     def populate_sheet_container(self) -> None:
-        # TODO: Simplify method
         if hasattr(self.sheet_container_box, "remove_all") and False:
             # Only available in GTK 4.12+
             self.sheet_container_box.remove_all()
@@ -501,62 +577,58 @@ class KeyhintWindow(Gtk.ApplicationWindow):
                 self.sheet_container_box.remove(child)
 
         sheet_id = self.headerbar.sheet_drop_down.get_selected_item().get_string()
-        sheet = self.get_sheet_by_id(sheet_id)
-
-        shortcut_column_factory = self.create_list_item_factory(
-            self.bind_shortcuts_callback
-        )
-        label_column_factory = self.create_list_item_factory(self.bind_labels_callback)
-
-        self.list_view_filter = Gtk.CustomFilter.new(
-            match_func=self.list_view_match_func
-        )
+        sheet = sheets.get_sheet_by_id(sheets=self.sheets, sheet_id=sheet_id)
 
         self.max_shortcut_width = 0
 
         sections = sheet["section"]
-        if not self.cli_args.get("no-section-sort", False):
-            # Sort sections by number of bindings to make the layout more dense
-            sections = dict(
-                sorted(sections.items(), key=lambda x: len(x[1]), reverse=True)
-            )
 
-        for section, bindings in sections.items():
-            ls = Gio.ListStore()
-            for shortcut, label in bindings.items():
-                ls.append(BindingsRow(shortcut=shortcut, label=label, section=section))
-
-            filter_list = Gtk.FilterListModel(model=ls)
-            filter_list.set_filter(self.list_view_filter)
-
-            selection = Gtk.NoSelection.new(filter_list)
-
-            # TODO: Dynamic width based on content
-            shortcut_column = self.create_column_view_column(
-                "", shortcut_column_factory, 220
-            )
-            label_column = self.create_column_view_column(section, label_column_factory)
-
-            column_view = self.create_column_view(
-                selection, shortcut_column, label_column
-            )
-
-            section_child = Gtk.FlowBoxChild()
-            section_child.set_vexpand(False)
-            section_child.set_child(column_view)
-
+        for index, (section, bindings) in enumerate(sections.items()):
+            section_child = self.create_section(section, bindings)
+            section_child.set_name(f"section-{index:03}")
             self.sheet_container_box.append(section_child)
 
-    def get_debug_info(self) -> str:
-        text = "Active Window:\n"
-        text += f"\ttitle: {self.window_title}\n"
-        text += f"\twmclass: {self.wm_class}\n"
-        text += "\nSelected Cheatsheet:\n"
-        text += f"\tID: {self.sheet_id}\n"
-        if sheet := self.get_sheet_by_id(self.sheet_id):
-            text += f"\tregex_process: {sheet['match']['regex_process']}\n"
-            text += f"\tregex_title: {sheet['match']['regex_title']}\n"
-            text += f"\tsource: {sheet['source']}\n"
-        else:
-            text += "No cheatsheet for this ID found!"
-        return text
+    def create_section(self, section: str, bindings: dict[str, str]) -> Gtk.Box:
+        ls = Gio.ListStore()
+        for shortcut, label in bindings.items():
+            ls.append(BindingsRow(shortcut=shortcut, label=label, section=section))
+
+        filter_list = Gtk.FilterListModel(model=ls)
+        filter_list.set_filter(self.list_view_filter)
+
+        selection = Gtk.NoSelection.new(filter_list)
+
+        # TODO: Dynamic width based on content
+        shortcut_column_width = self.config["main"].getint("zoom", 100) * 1.1 + 135
+        shortcut_column = utils.create_column_view_column(
+            "", self.shortcut_column_factory, shortcut_column_width
+        )
+        label_column = utils.create_column_view_column(
+            section, self.label_column_factory
+        )
+        column_view = utils.create_column_view(selection, shortcut_column, label_column)
+
+        section_child = Gtk.FlowBoxChild()
+        section_child.set_vexpand(False)
+        section_child.set_child(column_view)
+        return section_child
+
+    def get_debug_info_text(self) -> str:
+        sheet = sheets.get_sheet_by_id(sheets=self.sheets, sheet_id=self.sheet_id)
+        return (
+            "\n"
+            "<big>Last Active Application</big>\n\n"
+            f"<span foreground='#FF2E88'>title:</span> {self.window_title}\n"
+            f"<span foreground='#FF2E88'>wmclass:</span> {self.wm_class}\n"
+            "\n\n"
+            "<big>Selected Cheatsheet</big>\n\n"
+            f"<span foreground='#FF2E88'>ID:</span> {self.sheet_id}\n"
+            "<span foreground='#FF2E88'>regex_wmclass:</span> "
+            f"{sheet.get('match',{}).get('regex_wmclass', 'n/a')}\n"
+            "<span foreground='#FF2E88'>regex_title:</span> "
+            f"{sheet.get('match', {}).get('regex_title', 'n/a')}\n"
+            "<span foreground='#FF2E88'>source:</span> "
+            f"<a href='{sheet.get('url', '')}'>"
+            f"<span foreground='#00A6FF'>{sheet.get('url', 'n/a')}</span>"
+            "</a>\n"
+        )